Calculating function in another object
Better seperation of calculation and spawning of objects
